Renamed Crop.py to ImageTransform.py
